Revert "MWPW-117620 - Milo: Automated Breadcrumbs (CC/DC)" (#572)
Revert "MWPW-117620 - Milo: Automated Breadcrumbs (CC/DC) (#490)"

This reverts commit aa6e4e1e5003ce466defbe646f52dea5a95cff5f.
