+++ libs/blocks/gnav/gnav.js
@@ -13,8 +13,6 @@ import {
   analyticsGetLabel,
 } from '../../martech/attributes.js';
 
-import addBreadcrumbs from './gnav-breadcrumbs.js';
-
 const COMPANY_IMG = '<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 133.46 118.11"><defs><style>.cls-1{fill:#fa0f00;}</style></defs><polygon class="cls-1" points="84.13 0 133.46 0 133.46 118.11 84.13 0"/><polygon class="cls-1" points="49.37 0 0 0 0 118.11 49.37 0"/><polygon class="cls-1" points="66.75 43.53 98.18 118.11 77.58 118.11 68.18 94.36 45.18 94.36 66.75 43.53"/></svg>';
 const BRAND_IMG = '<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 240 234"><defs><style>.cls-1{fill:#fa0f00;}.cls-2{fill:#fff;}</style></defs><rect class="cls-1" width="240" height="234" rx="42.5"/><path id="_256" data-name="256" class="cls-2" d="M186.617,175.95037H158.11058a6.24325,6.24325,0,0,1-5.84652-3.76911L121.31715,99.82211a1.36371,1.36371,0,0,0-2.61145-.034l-19.286,45.94252A1.63479,1.63479,0,0,0,100.92626,148h21.1992a3.26957,3.26957,0,0,1,3.01052,1.99409l9.2814,20.65452a3.81249,3.81249,0,0,1-3.5078,5.30176H53.734a3.51828,3.51828,0,0,1-3.2129-4.90437L99.61068,54.14376A6.639,6.639,0,0,1,105.843,50h28.31354a6.6281,6.6281,0,0,1,6.23289,4.14376L189.81885,171.046A3.51717,3.51717,0,0,1,186.617,175.95037Z"/></svg>';
 const SEARCH_ICON = '<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" focusable="false"><path d="M14 2A8 8 0 0 0 7.4 14.5L2.4 19.4a1.5 1.5 0 0 0 2.1 2.1L9.5 16.6A8 8 0 1 0 14 2Zm0 14.1A6.1 6.1 0 1 1 20.1 10 6.1 6.1 0 0 1 14 16.1Z"></path></svg>';
@@ -53,7 +51,7 @@ class Gnav {
     this.desktop = window.matchMedia('(min-width: 1200px)');
   }
 
-  init = async () => {
+  init = () => {
     this.state = {};
     this.curtain = createTag('div', { class: 'gnav-curtain' });
     const nav = createTag('nav', { class: 'gnav', 'aria-label': 'Main' });
@@ -94,7 +92,11 @@ class Gnav {
 
     const wrapper = createTag('div', { class: 'gnav-wrapper' }, nav);
 
-    await addBreadcrumbs(this.el, wrapper);
+    this.setBreadcrumbSEO();
+    const breadcrumbs = this.decorateBreadcrumbs();
+    if (breadcrumbs) {
+      wrapper.append(breadcrumbs);
+    }
     decorateLinks(wrapper);
     this.el.append(this.curtain, wrapper);
   };
@@ -501,6 +503,40 @@ class Gnav {
       window.adobeIMS.signIn();
     });
     profileEl.append(signIn);
+  };
+
+  setBreadcrumbSEO = () => {
+    const seoEnabled = getMetadata('breadcrumb-seo') !== 'off';
+    if (!seoEnabled) return;
+    const breadcrumb = this.el.querySelector('.breadcrumbs');
+    if (!breadcrumb) return;
+    const breadcrumbSEO = { '@context': 'https://schema.org', '@type': 'BreadcrumbList', itemListElement: [] };
+    const items = breadcrumb.querySelectorAll('ul > li');
+    items.forEach((item, idx) => {
+      const link = item.querySelector('a');
+      breadcrumbSEO.itemListElement.push({
+        '@type': 'ListItem',
+        position: idx + 1,
+        name: link ? link.innerHTML : item.innerHTML,
+        item: link?.href,
+      });
+    });
+    const script = createTag('script', { type: 'application/ld+json' }, JSON.stringify(breadcrumbSEO));
+    document.head.append(script);
+  };
+
+  decorateBreadcrumbs = () => {
+    const parent = this.el.querySelector('.breadcrumbs');
+    if (parent) {
+      const ul = parent.querySelector('ul');
+      if (ul) {
+        ul.querySelector('li:last-of-type')?.setAttribute('aria-current', 'page');
+        const nav = createTag('nav', { class: 'breadcrumbs', 'aria-label': 'Breadcrumb' }, ul);
+        parent.remove();
+        return nav;
+      }
+    }
+    return null;
   };
   /* c8 ignore stop */
