--- test/blocks/gnav/mocks/breadcrumbs/from-file/breadcrumbs.plain.html
@@ -1,13 +1,0 @@
-<div>
-    <div class="breadcrumbs">
-        <div>
-            <div>
-                <ul>
-                    <li><a href="http://www.google.com/">from-file</a></li>
-                    <li><a href="http://www.adobe.com/">Actors</a></li>
-                    <li><a href="http://www.day.com">Players</a></li>
-                </ul>
-            </div>
-        </div>
-    </div>
-</div>
