--- test/blocks/gnav/mocks/breadcrumbs/from-url/level1/index.html
@@ -1,5 +1,0 @@
-<html>
-<head>
-    <title>level1 title</title>
-</head>
-</html>
