--- test/blocks/gnav/mocks/breadcrumbs/from-url/level1/level2/index.html
@@ -1,5 +1,0 @@
-<html>
-<head>
-    <title>level2 title</title>
-</head>
-</html>
