--- test/blocks/gnav/mocks/breadcrumbs/from-url/level1/level2/level3/index.html
@@ -1,5 +1,0 @@
-<html>
-<head>
-    <title>level3 title</title>
-</head>
-</html>
