--- test/blocks/gnav/mocks/breadcrumbs/hide-page/level1/index.html
@@ -1,6 +1,0 @@
-<html>
-<head>
-    <title>level1 title</title>
-    <meta name="breadcrumbs-hide-page" content="on">
-</head>
-</html>
